Feature fia register form (#20)
* moved loginForm components out of header folder, fixed pages folder name

* created registerForm, registerPage, imported LoginForm to LoginPage

* fixed scss misspell and links

* visibility switch to password

* deleted outer div on both Login and Register form
