+++ src/App.tsx
@@ -1,7 +1,9 @@
 import { Routes, BrowserRouter, Route } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
+import RegisterPage from "./pages/RegisterPage";
 import Layout from "./pages/Layout";
 import NotificationPage from "./pages/NotificationPage";
+
 function App() {
   return (
     <div>
@@ -10,6 +12,7 @@ function App() {
           <Route path="/" element={<Layout />} />
           <Route path="NotificationPage" element={<NotificationPage />} />
           <Route path="loginPage" element={<LoginPage />} />
+          <Route path="registerPage" element={<RegisterPage />} />
         </Routes>
       </BrowserRouter>
     </div>
