+++ src/Pages/RegisterPage.tsx
@@ -1,0 +1,7 @@
+import RegisterForm from "../components/RegisterForm/RegisterForm";
+
+function RegisterPage() {
+  return <RegisterForm />;
+}
+
+export default RegisterPage;
