+++ src/components/LoginForm/loginform.module.scss
@@ -1,7 +1,7 @@
 div {
   max-width: 466px;
 
-  & .orSignUp {
+  & .orRegister {
     &:hover {
       cursor: pointer;
     }
