+++ src/components/RegisterForm/registerform.module.scss
@@ -1,0 +1,17 @@
+div {
+    max-width: 466px;
+  
+    & .orLogin {
+      &:hover {
+        cursor: pointer;
+      }
+    }
+  
+    & .registerFormControl {
+      @media (min-width: 400px) {
+        width: 322px;
+      }
+    }
+    
+  }
+
